refactor: reorganize ids
