+++ Backend/src/BuildingBlocks/BuildingBlocks.Domain/ValueObjects/Ids/FileAssetId.cs
@@ -1,4 +1,6 @@
-﻿namespace BuildingBlocks.Domain.ValueObjects.Ids;
+﻿using BuildingBlocks.Domain.Abstractions;
+
+namespace BuildingBlocks.Domain.ValueObjects.Ids;
 
 public class FileAssetId : StronglyTypedId
 {
